refactor(ms-bing-ads-audiences): inline createAudience error handling, address review
- Per review feedback, keep the createAudience flow inline in index.ts (as it
  was originally) instead of extracting it to a createBingAudience helper —
  smaller diff, lower regression risk. Only the reusable readResponseBody helper
  stays in utils.ts.
- Default a missing/unknown upstream status to 500 (retryable) instead of 400, so
  ambiguous failures aren't misclassified as non-retryable client errors.
- Wrap non-HTTPError failures (timeouts, socket/connection errors) as a retryable
  IntegrationError instead of letting a raw opaque error escape.
- readResponseBody now handles non-string content (Buffer/object) before falling
  back to text().
- Add tests for the connection-failure path.
