+++ packages/destination-actions/src/destinations/ms-bing-ads-audiences/__tests__/index.test.ts
@@ -145,6 +145,15 @@ describe('Bing Ads Audiences', () => {
       expect(error.code).toBe('CREATE_AUDIENCE_FAILED')
       expect(error.status).toBe(500)
     })
+
+    it('should default to a retryable 500 status when the connection fails with no HTTP response', async () => {
+      nock(BASE_URL).post('/Audiences').replyWithError('socket hang up')
+
+      const error = await testDestination.createAudience(createAudienceInput).catch((e) => e)
+      expect(error).toBeInstanceOf(IntegrationError)
+      expect(error.code).toBe('CREATE_AUDIENCE_FAILED')
+      expect(error.status).toBe(500)
+    })
   })
 
   describe('getAudience', () => {
